+++ src/cmd/append.go
@@ -11,7 +11,7 @@ import (
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/validate"
 	"github.com/git-town/git-town/v9/src/vm/interpreter"
-	"github.com/git-town/git-town/v9/src/vm/state"
+	"github.com/git-town/git-town/v9/src/vm/runstate"
 	"github.com/spf13/cobra"
 )
 
@@ -58,7 +58,7 @@ func executeAppend(arg string, debug bool) error {
 	if err != nil || exit {
 		return err
 	}
-	runState := state.RunState{
+	runState := runstate.RunState{
 		Command:             "append",
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            appendSteps(config),
