+++ src/cmd/prepend.go
@@ -13,7 +13,7 @@ import (
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/validate"
 	"github.com/git-town/git-town/v9/src/vm/interpreter"
-	"github.com/git-town/git-town/v9/src/vm/state"
+	"github.com/git-town/git-town/v9/src/vm/runstate"
 	"github.com/spf13/cobra"
 )
 
@@ -61,7 +61,7 @@ func executePrepend(args []string, debug bool) error {
 	if err != nil || exit {
 		return err
 	}
-	runState := state.RunState{
+	runState := runstate.RunState{
 		Command:             "prepend",
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            prependSteps(config),
