+++ src/cmd/rename_branch.go
@@ -11,7 +11,7 @@ import (
 	"github.com/git-town/git-town/v9/src/step"
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/vm/interpreter"
-	"github.com/git-town/git-town/v9/src/vm/state"
+	"github.com/git-town/git-town/v9/src/vm/runstate"
 	"github.com/spf13/cobra"
 )
 
@@ -67,7 +67,7 @@ func executeRenameBranch(args []string, force, debug bool) error {
 	if err != nil || exit {
 		return err
 	}
-	runState := state.RunState{
+	runState := runstate.RunState{
 		Command:             "rename-branch",
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            renameBranchSteps(config),
