+++ src/cmd/ship.go
@@ -16,7 +16,7 @@ import (
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/validate"
 	"github.com/git-town/git-town/v9/src/vm/interpreter"
-	"github.com/git-town/git-town/v9/src/vm/state"
+	"github.com/git-town/git-town/v9/src/vm/runstate"
 	"github.com/spf13/cobra"
 )
 
@@ -91,7 +91,7 @@ func executeShip(args []string, message string, debug bool) error {
 			return err
 		}
 	}
-	runState := state.RunState{
+	runState := runstate.RunState{
 		Command:             "ship",
 		InitialActiveBranch: initialBranchesSnapshot.Active,
 		RunSteps:            shipSteps(config, message),
