+++ src/cmd/undo.go
@@ -13,7 +13,7 @@ import (
 	"github.com/git-town/git-town/v9/src/steps"
 	"github.com/git-town/git-town/v9/src/vm/interpreter"
 	"github.com/git-town/git-town/v9/src/vm/persistence"
-	"github.com/git-town/git-town/v9/src/vm/state"
+	"github.com/git-town/git-town/v9/src/vm/runstate"
 	"github.com/spf13/cobra"
 )
 
@@ -110,13 +110,13 @@ func determineUndoConfig(repo *execute.OpenRepoResult, debug bool) (*undoConfig,
 	}, initialStashSnapshot, lineage, nil
 }
 
-func determineUndoRunState(config *undoConfig, repo *execute.OpenRepoResult) (state.RunState, error) {
+func determineUndoRunState(config *undoConfig, repo *execute.OpenRepoResult) (runstate.RunState, error) {
 	runState, err := persistence.Load(repo.RootDir)
 	if err != nil {
-		return state.RunState{}, fmt.Errorf(messages.RunstateLoadProblem, err)
+		return runstate.RunState{}, fmt.Errorf(messages.RunstateLoadProblem, err)
 	}
 	if runState == nil || runState.IsUnfinished() {
-		return state.RunState{}, fmt.Errorf(messages.UndoNothingToDo)
+		return runstate.RunState{}, fmt.Errorf(messages.UndoNothingToDo)
 	}
 	undoRunState := runState.CreateUndoRunState()
 	undoRunState.RunSteps.Wrap(steps.WrapOptions{
