+++ src/vm/runstate/core.go
@@ -1,2 +1,2 @@
 // Package state represents the current execution state of the interpreter within a Git Town command.
-package state
+package runstate
