+++ src/vm/runstate/unfinished_state_details.go
@@ -1,4 +1,4 @@
-package state
+package runstate
 
 import (
 	"fmt"
